More 3.0 updates

+++ include/calc/Rle.hpp
@@ -27,7 +27,7 @@ public:
 	 * @param src_size The number of bytes to encode
 	 * @return Number of un-encoded bytes that were processed
 	 */
-	static int encode(void * dest, ssize_t & dest_size, const void * src, ssize_t src_size);
+	static int encode(void * dest, s32 & dest_size, const void * src, s32 src_size);
 
 	/*! \details Decodes a block of data that has been run length encoded.
 	 *
@@ -37,7 +37,7 @@ public:
 	 * @param src_size The number of encoded bytes to process
 	 * @return Number of encoded bytes that were processed
 	 */
-	static int decode(void * dest, ssize_t & dest_size, const void * src, ssize_t src_size);
+	static int decode(void * dest, s32 & dest_size, const void * src, s32 src_size);
 
 	/*! \details Calculates the number of bytes that will be used by the
 	 * compressed data.
